#107 [FEAT] 오버레이 드래그 훅 분리
